review

+++ include/librtmvideo/video_bot.h
@@ -3,11 +3,12 @@
 
 // Video bot API.
 extern "C" {
-using bot_callback_t = void (*)(const uint8_t *image, int width, int height);
+using bot_callback_t = void (*)(const uint8_t *image, uint16_t width,
+                                uint16_t height);
 
 struct bot_descriptor {
-  int image_width;
-  int image_height;
+  uint16_t image_width;
+  uint16_t image_height;
   bot_callback_t callback;
 };
